Refactoring tokenizers + adding OpenAI tokenizer (#489)
* separated out tokenizers

* separated out tokenizers

* fixed formatting

* reorganizing data loading

* small cleaning up changes

* added todo for detokenize

* cleaning up

* removed tests folder

* fixing tokenizer references
